+++ .eslintrc.js
@@ -3,29 +3,41 @@ rulesDirPlugin.RULES_DIR = "lib/rules";
 
 module.exports = {
   root: true,
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 2021,
     sourceType: "module",
     ecmaFeatures: {
       jsx: true,
-    },
+    }
   },
   extends: [
     "eslint:recommended",
     "plugin:eslint-plugin/recommended",
     "plugin:node/recommended",
+    "plugin:import/typescript"
   ],
   env: {
+    browser: true,
     node: true,
   },
-  plugins: ["rulesdir"],
+  plugins: ["rulesdir", "import"],
   rules: {
     "rulesdir/order": 2,
+    "import/no-unresolved": "off",
+    "node/no-missing-import": "off",
+    "node/no-unsupported-features/es-syntax": [
+      "error",
+      { ignores: ["modules"] },
+    ],
   },
-  overrides: [
-    {
-      files: ["tests/**/*.js"],
-      env: { mocha: true },
-    },
-  ],
+  ignorePatterns: ["playground/**/*"],
+  settings: {
+    'import/resolver': {
+      node: {
+        paths: ["src"],
+        extensions: ['.js', '.jsx', '.ts', '.tsx']
+      }
+    }
+  }
 };